feat(接口): 分数不做小数保存

+++ housekeeping-common/src/main/java/com/housekeeping/common/entity/ScoreCalculation.java
@@ -57,7 +57,7 @@ public class ScoreCalculation {
     private Float getScope2(){
         Float maxScope = new Float(weight.get(ApplicationConfigConstants.distanceScoreDouble));
         Float x = new Float(this.variable6);
-        Float y = (-1) * maxScope * x/new Float(50000) + maxScope;
+        Float y = (-1) * maxScope * x/new Float(50) + maxScope;
         return y;
     }
